Remove tests for legacy verifiers
